Switch from SweetXML to the official Google Cloud API library
There is only one problem with the current Google Cloud API library,
which is that the Cloud Storage API **DOES NOT** support uploading
binary data to create objects, you must always pass in a path to a file.
This is infuriating for many reasons, but chief among them for me is
because the underlying HTTP library (Tesla) uses the raw file contents
to create a mutli-part form request if you pass in a file path and not
the binary data, so we would actually be adding extra steps to the
process if we wrote binary data to a temporary file and used that.

Because of this, I needed to copy/paste the Google Cloud Storage code
for creating and executing an upload request and modify it so that I was
setting the request body in a specific manner that bypassed the special
field-setting code that Google uses.